backend/requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.8.0
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
fastapi==0.115.6
h11==0.14.0
idna==3.10
macholib==1.16.3
numpy==2.2.1
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
psutil==6.1.1
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32-ctypes==0.2.3
pyxlsb==1.0.10
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
